Fixed problem with MarshallTools for Marshall Entries with extra commas.
modified:   MarshalLightcurve.py
- Problem arises due to the presence of commas and un-paired quotation
marks in comments in the last but one column `refsys`.
- Wrapped in try/except loop for transients that have this problem
- Within these transients, use a line by line treatment, passing lines
with the correct column numbers
- For lines with higher column numbers delimted by commas, we assume all
of the extra columns should be in the `refsys` column, and fix it
replacing commas by semi-colons.
- We then use the same method of parsing strings as the rest for
uniformity, rather than levereging the parsing done.
- A problem that exists is that the comments don't have spaces between
words, but this seems to be introduced before the code doing  the baove
fix

modified:   ../examples/Demo_MarshalTools.ipynb
- Modified the Demo to include an example of the previously problematic
ones

modified:   version.py
- bump version